refactor: move transaction out of swap package
