+++ pkg/postage/postagecontract/contract_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	postageMock "github.com/ethersphere/bee/pkg/postage/mock"
 	"github.com/ethersphere/bee/pkg/postage/postagecontract"
-	"github.com/ethersphere/bee/pkg/settlement/swap/transaction"
-	transactionMock "github.com/ethersphere/bee/pkg/settlement/swap/transaction/mock"
+	"github.com/ethersphere/bee/pkg/transaction"
+	transactionMock "github.com/ethersphere/bee/pkg/transaction/mock"
 )
 
 func TestCreateBatch(t *testing.T) {
