+++ pkg/settlement/swap/chequebook/chequebook_test.go
@@ -15,9 +15,9 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethersphere/bee/pkg/settlement/swap/chequebook"
 	erc20mock "github.com/ethersphere/bee/pkg/settlement/swap/erc20/mock"
-	"github.com/ethersphere/bee/pkg/settlement/swap/transaction"
-	transactionmock "github.com/ethersphere/bee/pkg/settlement/swap/transaction/mock"
 	storemock "github.com/ethersphere/bee/pkg/statestore/mock"
+	"github.com/ethersphere/bee/pkg/transaction"
+	transactionmock "github.com/ethersphere/bee/pkg/transaction/mock"
 )
 
 func TestChequebookAddress(t *testing.T) {
